+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 All notable changes to this project should be documented in this file.
 
+### v0.0.6
+- support multiple telegram channel
+
 ### v0.0.5
-- support for multiple telegram group
-- support for multiple twitter account
+- support multiple telegram group
+- support multiple twitter account
 - refactor configuration to file based configuration
 
 ### v0.0.4